Set "enabled" and "updated"
If not set in the request, "enabled" is set to false, as this is the
default value for boolean variables. To fix this "enabled" type is
changed to Boolean and if not null (set in the request) the value is
updated.

The "updated" property is only set when the entity is created and
subsequent updates didn't change its value.